refactor(org):split org controller into 2
